refactor: restructure CowSaucerScene for improved modularity
