+++ libs/ui/src/index.ts
@@ -40,6 +40,7 @@ export * from './pure/Input'
 export * from './pure/Badge'
 export * from './pure/NetworkLogo'
 export * from './pure/CircleProgress'
+export * from './pure/CowSaucerScene'
 
 export * from './containers/CowSwapSafeAppLink'
 export * from './pure/InlineBanner'
