+++ libs/ui/src/pure/CowSaucerScene/index.ts
@@ -1,0 +1,7 @@
+export { CowSaucerScene } from './CowSaucerScene'
+export type { CowSaucerSceneProps, SaucerPalette } from './CowSaucerScene'
+export {
+  SWAP_SAUCER_PALETTE_LIGHT,
+  SWAP_SAUCER_PALETTE_DARK,
+  COWFI_SAUCER_PALETTE_LIGHT,
+} from './palettes'
